+++ src/de/falk/battleships/configuration/ShotStrategy.scala
@@ -1,7 +1,6 @@
 package de.falk.battleships.configuration
 
-import de.falk.battleships.model.Coordinate
-import de.falk.battleships.model.InformationForShot
+import de.falk.battleships.model.{Coordinate, ShipType}
 
 sealed trait ShotStrategy
 
@@ -10,7 +9,7 @@ case object OnlyRandomOfNotShot extends ShotStrategy
 case object RandomOfNotShotOrAroundPartialHit extends ShotStrategy
 
 object ShotStrategy {
-   def onlyRandomOfNotShot(informationForShot: InformationForShot) : Coordinate = {
-    scala.util.Random.shuffle(informationForShot.notShots).head
+   def onlyRandomOfNotShot(informationForShot: (List[Coordinate], List[(Coordinate, ShipType)])) : Coordinate = {
+    scala.util.Random.shuffle(informationForShot._1).head
   }
 }