+++ src/de/falk/battleships/model/Types.scala
@@ -1,5 +1,0 @@
-package de.falk.battleships.model
-
-case class InformationForShot(notShots: List[Coordinate], partialHits: List[(Coordinate, ShipType)])
-
-case class ShotResult(simpleOrComplex: ShotResultValue)